+++ libs/src/ChangeLog.md
@@ -6,10 +6,12 @@
 
 🆙update
 
+- 帶星號元件 story 名稱改全中文，第一個皆改預設樣式
+
 - 修改 checkbox vue 元件寫法
 - 修改 checkbox story code 顯示方式
 - 修改 checkbox story table 顯示方式
-- 帶星號元件 story 名稱改全中文，第一個皆改預設樣式
+
 - 修改 breadcrumb vue 元件寫法，teleport body 計算定位位置
 - 修改 breadcrumb story code 顯示方式
 - 修改 breadcrumb story table 顯示方式
@@ -17,6 +19,7 @@
 - 新增 Avatar parameters.docs.source 作為下載程式碼用
 - 新增 AvatarGroup parameters.docs.source 作為下載程式碼用
 - 新增 Badge parameters.docs.source 作為下載程式碼用
+- 新增 Button parameters.docs.source 作為下載程式碼用
 
 - 新增 AvatarGroup 必填項目
 - 新增 Badge 必填項目